BUG: Fix Conversion::ToString(float) itk::NumberToString specialization
Accidentally, `ToString(float)` called the "double" specialization of `itk::NumberToString`, which led to a sub-optimal string representation. For example, 0.1f would yield "0.10000000149011612", instead of simply "0.1".

Fixed by calling the "float" specialization of `itk::NumberToString` instead.

+++ Core/Install/elxConversion.cxx
@@ -161,7 +161,7 @@ Conversion::ToString(const double scalar)
 std::string
 Conversion::ToString(const float scalar)
 {
-  return itk::NumberToString<double>{}(scalar);
+  return itk::NumberToString<float>{}(scalar);
 }
